requirements/base.txt
Django==4.1.3  # Programming framework - https://github.com/django/django
django-cors-headers==3.13.0  # Allowing domains/regix for CORS - https://github.com/adamchainz/django-cors-headers
django-environ==0.9.0  # Use .env for environment variables - https://github.com/joke2k/django-environ
django-import-export==3.0.1  # Import/export in django admin - https://github.com/django-import-export/django-import-export
Pillow==9.2.0  # Image modifying - https://github.com/python-pillow/Pillow
whitenoise==6.2.0  # Simplified static file serving - https://github.com/evansd/whitenoise
sentry-sdk==1.10.1  # https://github.com/getsentry/sentry-python
django_health_check==3.17.0  # Status page - https://github.com/KristianOellegaard/django-health-check
GitPython==3.1.29  # Git integration for versioning - https://gitpython.readthedocs.io/en/stable/
celery==5.2.7
channels-redis==3.4.1
channels==3.0.5
django-graphql-jwt==0.3.4
django-redis==5.2.0
django-rq==2.6.0
graphene==3.1
graphene-django==3.0.0
graphene-django-extras==0.5.2
graphene-file-upload==1.3.0
mysqlclient==2.1.1
python-graphql-client==0.4.3
django-graphql-playground==1.0.0
martor==1.6.14
django-admin-sortable2==2.1.4
django-mptt==0.14.0
django-modeladmin-reorder==0.3.1
django-storages==1.11.1
boto3==1.26.13
deepl==1.11.0
celery_once==3.0.1